add ReadHandlerMap
Supports generating a map of read handlers once and only once, and then
passing that to TransitFactory.reader(format, in, handlers). When the
reader method sees that handlers is a ReadHandlerMap, it just uses it
instead of looking it up in the cache and then, if not cached, building
it.